backport CVS commit by pfeiffer:
The sequence of connection types in preferences->connection->connectiontype
and in settings.cpp line 47 are not in tune. Thus, when I choose connection type
"PPP", kget checks for a "PLIP" connection and never notices when I'm online.
The patch changes the order of connection types as defined in dlgconnectionbase.ui

svn path=/branches/KDE_3_2_BRANCH/kdenetwork/kget/; revision=289000