KGet has now a multithread http/ftp transfer plugin! It's in development, but it works in basic.
It's written by Manolo Valdes (Manolito / nolis71cu).

In order to resume a FTP transfer, the plugin depends on a patched qt-copy. (but this is disabled by default, so it'll compile with our qt-copy)
Manolito has already sent the patch to Trolltech. Let's hope it'll be in Qt4.2.

CCMAIL: nolis71cu@gmail.com

svn path=/branches/work/make_kget_cool/kget/; revision=506513