Backport r1072308
Removes the seperation of addSegment and addSegments, now one can only use addSegments. The TransferDataSources have to handle all that internally. Also removes the "offest" in addSegments, the offset has to be calculated by the TransferDataSources (segmentRange.first + segmentSize.first), the reason is to ensure that offset is always based on segmentRange and segmentSize and nothing else.

svn path=/branches/KDE/4.4/kdenetwork/kget/; revision=1072309