Adds FileDeleter which manages the removal of files.
Users of FileDeleter simply specify the url of the file
to delete and are then able to connect to the job.
FileDeleter makes sure that for each file there is only
one job, i.e. it is possible to call deleteFile multiple
times.
Also adds a unit test for FileDeleter.

svn path=/trunk/KDE/kdenetwork/kget/; revision=1280602