Return the (Empty) List of Transfers when File Name is Not Provided
The code wasn't really going to do anything with it later, anyways:
control goes through the code that handle multiple downloads. The UX
that this causes is that, when you cancel a download added through DBus
(say, using your browser), then you are asked to provide a directory
and that is another dialog to cancel.

REVIEW: 113053