Keep Header Menu Synced with Header Columns
Once the header menu was initialized it would never be updated with the
columns added to the menu. You could see this by right-clicking on the
headers and noticing that the menu only had a title and no columns to
check/uncheck. This change keeps the header menu updated every time the
header changes its column count.

BUG: 329957
FIXED-IN: 4.12.2
REVIEW: 115320