Fix torrent download with remote .torrent files
In case of a remote Url, the .torrent file is downloaded to a temporary
location first, and then that local file is used as source for the
actual torrent download.
But QUrl::QUrl() (that's used implicitly here to convert the QString to
QUrl) expects an Url, passing it a local path (without "file://")
results in an invalid Url which caused the download to fail.

Using QUrl::fromLocalFile() to convert the local path to QUrl fixes that
problem.