Fix changing the destination in the "New Download" dialog
Amongst other things, checkInput() sets the destination according to the
value in the destination text field.
It is supposed to be called whenever the text in some input field
changes, but this doesn't seem to work for the destination field for
some reason.
So as a workaround it is called explicitly now when the dialog is
accepted.

This fixes the problem that downloads always ended up in the default
folder, the entered/selected destination was ignored.