Fix MMS downloads
MmsDownload() expects a path as destination.
QUrl::toString() returns the full Url as QString though, so the
downloaded data was saved to "file:///..." which failed.

Also, KIO:del() takes a QUrl as parameter, passing it a path (as
QString) results in an invalid QUrl (because of the implicit conversion)
and so the (temporary) files were not deleted.