Fix Transfer::directory()
This was broken in two ways:
- QUrl::path() returns the full path including the filename, not just
  the directory
- the implicit conversion of a path results in an invalid QUrl

This caused a glitch in metalink's file selection dialog that displayed
the full destination folder hierarchy of the files to download instead
of the relative path, as it removes the result of this function from
the destination filepaths.

The KDE4 code called KUrl::upUrl() here, so return the result of
KIO::upUrl() now. (although it probably would be sufficient here to just
remove the filename...)