Simplify and fix "Download again?" dialog
Call QDialog::done() in the button handlers to actually close the
dialog, and pass it the proper return code.
Get the result from dialog->exec() directly instead of calling result()
afterwards.

This eliminates the need for the m_result member variable and result()
function.

Also fix the check for the "Apply to All" checkbox state.