Port to ecm_add_test to actually run the tests
Requiring at least ECM 5.38 comes with a different build dir layout to
help testing of not installed apps. But in combination with the old
add_test syntax ctest can't find the test executable and thus can't run
the tests. Using ecm_add_test fixes that. The tests are executed with this
change, schedulertest does not pass here, though.