Fix destination filesystem type check for downloads bigger than 4 GiB
`m_dest` is a QUrl, we need to use `toLocalFile()` to get a local path.
`toString()` results in a URL instead ("file://...") which causes the
mountpoint lookup to fail.

Also, the check whether the mountpoint was found was reversed. It
actually tried to compare the filesystem type when the mount point was
*not* found, resulting in a crash because of the nullptr dereference.

BUG: 444591
FIXED-IN: 21.12.1