Fix build with recent KCoreAddons
In KCoreAddons the kcoreaddons_add_plugin macro doesn't add a plugins/
suffix if the required KF version >= 90 (really 38) see [1] for details.

This broke the build here since the min. required KF version was bumped but
the CMake code wasn't changed (plugins were put into builddir/kget/ and the
build system couldn't write the binary 'kget' in the builddir because a dir
with the same name already existed).

[1] https://invent.kde.org/frameworks/kcoreaddons/-/commit/375e19cfb